# openapi/v1.yaml
openapi: 3.1.0
info:
  title: En stor stark API
  version: 1.0.0
  summary: JSON API for the native En stor stark clients.
  description: |
    Readers use the public endpoints without authentication. Reviewers sign in with
    `POST /sessions` and send the returned token as `Authorization: Bearer <token>`.

    The API never reads the web session cookie, so the SvelteKit CSRF origin check does not
    apply to API clients. All requests and responses are JSON, except image downloads.

    User-facing messages (`detail`, `errors[].detail`) are Swedish and are safe to show to
    the user. Clients branch on `code` and HTTP status, never on message text.

    Public collection and detail responses include an `ETag`; send `If-None-Match` to get
    `304 Not Modified`. Review updates require `If-Match` with the review's current `ETag`.
servers:
  - url: /api/v1
...
security:
  - {}
  - bearerAuth: []
paths:
...
  /reviews:
    get:
      tags: [Reviews]
      operationId: listReviews
      summary: List reviews.
      description: |
        Anonymous clients get only published reviews. Authenticated clients also get drafts.
        The search matches the title, location, description, beer brand and authors.
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 80
        - name: sort
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ReviewSort'
      responses:
        '200':
          description: Matching reviews in the requested order.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
        '304':
          $ref: '#/components/responses/NotModified'
    post:
      tags: [Reviews]
      operationId: createReview
      summary: Create a draft review.
      description: New reviews are always drafts. Publish them with the publication endpoint.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
      responses:
        '201':
          description: Draft created.
          headers:
            Location:
              description: API path of the new review.
              schema:
                type: string
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The `token` from `POST /sessions`.
...
  headers:
    ETag:
      description: Entity tag of the response. Send it back in `If-None-Match` or `If-Match`.
      schema:
        type: string
...
  responses:
    NotModified:
      description: The cached representation is still current.
    BadRequest:
      description: The body is not valid JSON.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication is missing, invalid or expired.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: |
        The request conflicts with the current state. `code` is `duplicate_slug`,
        `already_published` or `concurrent_update`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    PayloadTooLarge:
      description: The request body or image is too large.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnsupportedMediaType:
      description: The request body is not `application/json`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationFailed:
      description: One or more fields are invalid. See `errors`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InternalError:
      description: The server could not complete the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details with a machine-readable `code`.
      required: [status, code, title, detail]
      properties:
        status:
          type: integer
        code:
          $ref: '#/components/schemas/ProblemCode'
        title:
          type: string
          description: Short English summary for logs.
        detail:
          type: string
          description: Swedish message that clients can show.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    ProblemCode:
      type: string
      enum:
        - bad_request
        - unauthorized
        - invalid_credentials
        - not_found
        - validation_failed
        - duplicate_slug
        - already_published
        - concurrent_update
        - precondition_failed
        - precondition_required
        - payload_too_large
        - unsupported_media_type
        - rate_limited
        - service_unavailable
        - internal_error
    FieldError:
      type: object
      required: [pointer, detail]
      properties:
        pointer:
          type: string
          description: JSON Pointer (RFC 6901) into the request body. Empty for the whole body.
          examples: ['/title', '/beer/priceKr', '/ratings/atmosphere']
        detail:
          type: string
          description: Swedish message that clients can show next to the field.
...
    Ratings:
      type: object
      description: Aspect ratings from 0 to 5.
      additionalProperties: false
      required:
        - atmosphere
        - service
        - selection
        - quality
        - price
        - cleanliness
        - soundLevel
        - barhopPotential
      properties:
        atmosphere: { type: number, minimum: 0, maximum: 5 }
        service: { type: number, minimum: 0, maximum: 5 }
        selection: { type: number, minimum: 0, maximum: 5 }
        quality: { type: number, minimum: 0, maximum: 5 }
        price: { type: number, minimum: 0, maximum: 5 }
        cleanliness: { type: number, minimum: 0, maximum: 5 }
        soundLevel: { type: number, minimum: 0, maximum: 5 }
        barhopPotential: { type: number, minimum: 0, maximum: 5 }
    RatingsInput:
      type: object
      description: Whole-number aspect ratings from 0 to 5.
      additionalProperties: false
      required:
        - atmosphere
        - service
        - selection
        - quality
        - price
        - cleanliness
        - soundLevel
        - barhopPotential
      properties:
        atmosphere: { type: integer, minimum: 0, maximum: 5 }
        service: { type: integer, minimum: 0, maximum: 5 }
        selection: { type: integer, minimum: 0, maximum: 5 }
        quality: { type: integer, minimum: 0, maximum: 5 }
        price: { type: integer, minimum: 0, maximum: 5 }
        cleanliness: { type: integer, minimum: 0, maximum: 5 }
        soundLevel: { type: integer, minimum: 0, maximum: 5 }
        barhopPotential: { type: integer, minimum: 0, maximum: 5 }
...
    ReviewSort:
      type: string
      enum: [latest, oldest, score]
      default: latest
    PublicationStatus:
      type: string
      enum: [draft, published]
    ReviewImage:
      type: object
      required: [url, focusX, focusY]
      properties:
        url:
          type: string
          description: Path relative to the server origin, for example `/api/v1/images/<file>`.
        focusX:
          type: number
          minimum: 0
          maximum: 100
          description: Horizontal focus point in percent for cropping.
        focusY:
          type: number
          minimum: 0
          maximum: 100
          description: Vertical focus point in percent for cropping.
    Beer:
      type: object
      required: [brand, priceKr, isHappyHourPrice]
      properties:
        brand:
          type: [string, 'null']
          description: Null for older reviews without a brand.
        priceKr:
          type: [integer, 'null']
          description: Null for older reviews without a valid price.
        isHappyHourPrice:
          type: boolean
    Review:
      type: object
      required:
        - id
        - slug
        - title
        - description
        - location
        - ratings
        - overallRating
        - beer
        - image
        - author
        - coAuthors
        - publicationStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
          description: Markdown.
        location:
          type: string
          description: Street address.
        ratings:
          $ref: '#/components/schemas/Ratings'
        overallRating:
          type: integer
          minimum: 0
          maximum: 3
        beer:
          $ref: '#/components/schemas/Beer'
        image:
          $ref: '#/components/schemas/ReviewImage'
        author:
          type: string
          description: Primary author.
        coAuthors:
          type: array
          items:
            type: string
        publicationStatus:
          $ref: '#/components/schemas/PublicationStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReviewList:
      type: object
      required: [reviews]
      properties:
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
    ImageContentType:
      type: string
      enum: [image/jpeg, image/png, image/webp]
    ImageUpload:
      type: object
      additionalProperties: false
      required: [contentType, data]
      properties:
        contentType:
          $ref: '#/components/schemas/ImageContentType'
        data:
          type: string
          contentEncoding: base64
          description: Standard base64 without a data-URL prefix. Convert HEIC to JPEG first.
    ImageFocus:
      type: object
      additionalProperties: false
      required: [x, y]
      properties:
        x:
          type: number
          minimum: 0
          maximum: 100
        y:
          type: number
          minimum: 0
          maximum: 100
    BeerInput:
      type: object
      additionalProperties: false
      required: [brand, priceKr, isHappyHourPrice]
      properties:
        brand:
          type: string
          minLength: 1
          description: A suggested brand from the metadata or another brand name.
        priceKr:
          type: integer
        isHappyHourPrice:
          type: boolean
    ReviewFields:
      type: object
      required: [title, description, location, ratings, beer, authors, imageFocus]
      properties:
        title:
          type: string
        description:
          type: string
          description: Markdown.
        location:
          type: string
        slug:
          type: string
          description: Optional. The server derives a slug from the title when it is omitted.
        ratings:
          $ref: '#/components/schemas/RatingsInput'
        overallRating:
          type: integer
          minimum: 0
          maximum: 3
          description: Optional. The server uses the suggested overall rating when it is omitted.
        beer:
          $ref: '#/components/schemas/BeerInput'
        authors:
          type: array
          description: |
            Credited reviewers. The signed-in user becomes the primary author when selected;
            otherwise an existing primary author, then the first selection.
          minItems: 1
          items:
            type: string
        imageFocus:
          $ref: '#/components/schemas/ImageFocus'
    ReviewCreateRequest:
      allOf:
        - $ref: '#/components/schemas/ReviewFields'
        - type: object
          required: [image]
          properties:
            image:
              $ref: '#/components/schemas/ImageUpload'
      unevaluatedProperties: false
...